+++ eng/tools/generator/cmd/v2/common/testdata/old/interfacetoany/interfacetoanyfilter.go
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+package interfacetoany
+
+type Interface2Any struct {
+	StringType interface{}
+	ArrayType  []interface{}
+	MapType    map[string]interface{}
+
+	NewType interface{}
+}